requirements.txt
click>=8.0
openai>=1.0
anthropic>=0.40
python-frontmatter>=1.0
lancedb>=0.6
networkx>=3.0
python-louvain>=0.16
jinja2>=3.0
streamlit>=1.30
pyyaml>=6.0